+++ htsinfer/adapters_list.txt
@@ -10,7 +10,6 @@ CCGGTTCTTCCCTGC
 CGACAGGTTCAGAGT
 CGGTCTCGGCATTCC
 CTGTAGGCACCATCA
-GAAUUCCACCACGUU
 GATCGGAAGAGCACA
 GATCGGAAGAGCGGT
 GATCGGAAGAGCTCG
@@ -18,11 +17,8 @@ GATCGTCGGACTGTA
 GCCTTGGCACCCGAG
 GTCTCGTGGGCTCGG
 GTGACTGGAGTTCAG
-GUUCAGAGUUCUACA
 TCGGACTGTAGAACT
 TCGTATGCCGTCTTC
 TCGTCGGCAGCGTCA
 TGGAATTCTCGGGTG
-UCGUAUGCCGUCUUC
 AGATCGGAAGAGCGT
-
